+++ jitsi-meet-nomad/jitsi-meet-nomad.d/local/share/cook/bin/configure-jitsi-videobridge.sh
@@ -31,15 +31,57 @@ sep=$'\001'
 # copy over sip-communicator.properties
 < "$TEMPLATEPATH/sip-communicator.properties.in" \
   sed "s${sep}%%publicip%%${sep}$PUBLICIP${sep}g" | \
-  sed "s${sep}%%privateip%%${sep}$IP${sep}g" | \
+  sed "s${sep}%%privateip%%${sep}$PRIVATEIP${sep}g" | \
   sed "s${sep}%%keypassword%%${sep}$KEYPASSWORD${sep}g" \
   > /usr/local/etc/jitsi/videobridge/sip-communicator.properties
+
+# if no image filename been passed in for a file copied-in to
+# /usr/local/www/jitsi-meet/image/filename.jpg then default to
+# watermark.svg
+# if an image has been passed, check for URL passed in
+if [ -n "$IMAGE" ]; then
+    IMAGE="$IMAGE"
+    export IMAGE
+    if [ -n "$LINK" ]; then
+        LINK="$LINK"
+        export LINK
+    fi
+else
+    IMAGE="watermark.svg"
+    export IMAGE
+fi
+
+# check if resolution has been passed in or set a default of 360
+if [ -n "$RESOLUTION" ]; then
+    RESOLUTION="$RESOLUTION"
+    export RESOLUTION
+else
+    RESOLUTION="360"
+    export RESOLUTION
+fi
 
 # copy over config.js
 < "$TEMPLATEPATH/config.js.in" \
   sed "s${sep}%%domain%%${sep}$DOMAIN${sep}g" | \
-  sed "s${sep}%%keypassword%%${sep}$KEYPASSWORD${sep}g" \
+  sed "s${sep}%%keypassword%%${sep}$KEYPASSWORD${sep}g" | \
+  sed "s${sep}%%resolution%%${sep}$RESOLUTION${sep}g" | \
+  sed "s${sep}%%image%%${sep}$IMAGE${sep}g" | \
+  sed "s${sep}%%link%%${sep}$LINK${sep}g" \
   > /usr/local/www/jitsi-meet/config.js
+
+# copy over interface_config.js
+< "$TEMPLATEPATH/interface_config.js.in" \
+  sed "s${sep}%%image%%${sep}$IMAGE${sep}g" | \
+  sed "s${sep}%%link%%${sep}$LINK${sep}g" \
+  > /usr/local/www/jitsi-meet/interface_config.js
+
+# copy over manifest.json
+cp -f "$TEMPLATEPATH/manifest.json.in" /usr/local/www/jitsi-meet/manifest.json
+
+# enable service
+sysrc jitsi_videobridge_flags="--apis=rest,xmpp" || true
+sysrc jitsi_videobridge_maxmem="3072m" || true
+service jitsi-videobridge enable || true
 
 # enable service
 sysrc jitsi_videobridge_flags="--apis=rest,xmpp"